requirements.txt
Django~=5.1.1
django-sesame
django-widget-tweaks
django-recaptcha
django-ses[events]
python-dotenv
